fix test_publish and remove merge mess
